+++ src/main/java/com/telepathicgrunt/the_bumblezone/mixin/blocks/BeeHiveDropsMixin.java
@@ -1,0 +1,26 @@
+package com.telepathicgrunt.the_bumblezone.mixin.blocks;
+
+import com.telepathicgrunt.the_bumblezone.enchantments.CombCutterEnchantment;
+import net.minecraft.block.BeehiveBlock;
+import net.minecraft.block.BlockState;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.ActionResultType;
+import net.minecraft.util.Hand;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.BlockRayTraceResult;
+import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(BeehiveBlock.class)
+public class BeeHiveDropsMixin {
+
+    // makes Comb Cutter increase drops from BeehiveBlocks.
+    @Inject(method = "onBlockActivated(Lnet/minecraft/block/BlockState;Lnet/minecraft/world/World;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/entity/player/PlayerEntity;Lnet/minecraft/util/Hand;Lnet/minecraft/util/math/BlockRayTraceResult;)Lnet/minecraft/util/ActionResultType;",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/block/BeehiveBlock;dropHoneyComb(Lnet/minecraft/world/World;Lnet/minecraft/util/math/BlockPos;)V"))
+    private void blockBreakingSpeed(BlockState state, World worldIn, BlockPos pos, PlayerEntity player, Hand handIn, BlockRayTraceResult hit, CallbackInfoReturnable<ActionResultType> cir) {
+        CombCutterEnchantment.increasedCombDrops(player, worldIn, pos);
+    }
+}