+++ src/main/java/com/telepathicgrunt/the_bumblezone/modcompat/CharmCompat.java
@@ -1,33 +1,8 @@
 package com.telepathicgrunt.the_bumblezone.modcompat;
 
-import com.minecraftabnormals.buzzier_bees.core.registry.BBBlocks;
-import com.minecraftabnormals.buzzier_bees.core.registry.BBItems;
-import com.telepathicgrunt.the_bumblezone.Bumblezone;
-import com.telepathicgrunt.the_bumblezone.mixin.blocks.DispenserBlockInvoker;
-import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.DispenserBlock;
-import net.minecraft.dispenser.DefaultDispenseItemBehavior;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-import net.minecraftforge.registries.ForgeRegistries;
-import svenhjol.charm.Charm;
 import svenhjol.charm.block.CandleBlock;
-import svenhjol.charm.module.BlockOfSugar;
 import svenhjol.charm.module.Candles;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Random;
 
 public class CharmCompat {