+++ src/main/java/com/telepathicgrunt/the_bumblezone/modcompat/ResourcefulBeesCompat.java
@@ -8,8 +8,8 @@ import com.resourcefulbees.resourcefulbees.block.multiblocks.apiary.ApiaryStorag
 import com.resourcefulbees.resourcefulbees.registry.BeeRegistry;
 import com.resourcefulbees.resourcefulbees.registry.ModBlocks;
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
-import com.telepathicgrunt.the_bumblezone.tags.BZBlockTags;
 import com.telepathicgrunt.the_bumblezone.modinit.BzConfiguredFeatures;
+import com.telepathicgrunt.the_bumblezone.tags.BZBlockTags;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;