+++ src/main/java/com/telepathicgrunt/the_bumblezone/world/processors/SpiderInfestedBeeDungeonProcessor.java
@@ -2,7 +2,6 @@ package com.telepathicgrunt.the_bumblezone.world.processors;
 
 import com.mojang.serialization.Codec;
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
-import com.telepathicgrunt.the_bumblezone.blocks.HoneyCrystal;
 import com.telepathicgrunt.the_bumblezone.blocks.HoneycombBrood;
 import com.telepathicgrunt.the_bumblezone.modcompat.*;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;