+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Enter the dimension by throwing an enderpearl at Bee Nest and exit it by going a
    Major:
 - Now requires 36.0.42 Forge or newer as I am now using EntityAttributeCreationEvent to register entity attributes.
 
+  Enchantments:
+- Added Comb Cutter enchantment for Shears and Swords. This will make mining all blocks with "comb" in the name much faster!
+  It will also increase the number of combs you get when shearing a vanilla Bee Hive / Bee Nest!
+
    Entities:
 - Using Honey Blocks on vanilla Slime mobs that are size 1 or 2 will turn them into Honey Slime mobs!
   "the_bumblezone:turn_slime_to_honey_slime" item tag controls what item can do the conversion.